requirements.txt
django==4.2
djangorestframework
python-magic==0.4.24